+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/json_streamer.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'json_streamer' => [
-        'enabled' => true,
-    ],
-]);