+++ src/Symfony/Component/JsonStreamer/Tests/DependencyInjection/CheckTypeInfoPassTest.php
@@ -9,17 +9,17 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\Compiler;
+namespace Symfony\Component\JsonStreamer\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\CheckJsonStreamerTypeInfoPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\RuntimeException;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\JsonStreamer\DependencyInjection\CheckTypeInfoPass;
 use Symfony\Component\JsonStreamer\JsonStreamWriter;
 use Symfony\Component\TypeInfo\TypeResolver\TypeResolver;
 
-class CheckJsonStreamerTypeInfoPassTest extends TestCase
+class CheckTypeInfoPassTest extends TestCase
 {
     public function testNothingIsAddedWhenTypeInfoIsAvailable()
     {
@@ -27,7 +27,7 @@ class CheckJsonStreamerTypeInfoPassTest extends TestCase
         $container->register('json_streamer.stream_writer', JsonStreamWriter::class);
         $container->register('type_info.resolver', TypeResolver::class);
 
-        new CheckJsonStreamerTypeInfoPass()->process($container);
+        new CheckTypeInfoPass()->process($container);
 
         $this->assertFalse($container->getDefinition('type_info.resolver')->hasErrors());
         $this->assertFalse($container->hasDefinition('type_info.type_context_factory'));
@@ -37,7 +37,7 @@ class CheckJsonStreamerTypeInfoPassTest extends TestCase
     {
         $container = new ContainerBuilder();
 
-        new CheckJsonStreamerTypeInfoPass()->process($container);
+        new CheckTypeInfoPass()->process($container);
 
         $this->assertFalse($container->hasDefinition('type_info.resolver'));
     }
@@ -47,7 +47,7 @@ class CheckJsonStreamerTypeInfoPassTest extends TestCase
         $container = new ContainerBuilder();
         $container->register('json_streamer.stream_writer', JsonStreamWriter::class);
 
-        new CheckJsonStreamerTypeInfoPass()->process($container);
+        new CheckTypeInfoPass()->process($container);
 
         foreach (['type_info.resolver', 'type_info.type_context_factory'] as $id) {
             $this->assertSame(['JsonStreamer support cannot be enabled as the TypeInfo component is not enabled. Try setting "type_info.enabled" to true.'], $container->getDefinition($id)->getErrors());
@@ -60,7 +60,7 @@ class CheckJsonStreamerTypeInfoPassTest extends TestCase
         $container->register('json_streamer.stream_writer', JsonStreamWriter::class)
             ->setPublic(true)
             ->setArguments([new Reference('type_info.resolver')]);
-        $container->addCompilerPass(new CheckJsonStreamerTypeInfoPass());
+        $container->addCompilerPass(new CheckTypeInfoPass());
 
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage('JsonStreamer support cannot be enabled as the TypeInfo component is not enabled. Try setting "type_info.enabled" to true.');