+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -24,6 +24,8 @@ CHANGELOG
  * Register `UidBundle` automatically when the Uid component is installed
  * Make `framework.scheduler` an alias of the `scheduler` configuration provided by `SchedulerBundle`
  * Register `SchedulerBundle` automatically when the Scheduler component is installed
+ * Make `framework.json_streamer` an alias of the `json_streamer` configuration provided by `JsonStreamerBundle`
+ * Register `JsonStreamerBundle` automatically when the JsonStreamer component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
  * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`